Time each iteration to minimise impact of GC on benchmarks
Prior to this commit, the timing included all activity from before
the first iteration to right after the last iteration including
any GC time and other time spent between `_run_iteration` behavior
calls.

This commit changes the logic to time each iteration individually
and accumulate the times of all iterations to avoid including any
time spent on GC and other things for synchronous benchmarks. A
similar change was made to asynchronous benchmarks, however those
are not guaranteed to exclude GC time due to the fact that the
actual async benchmark logic is likely to involve other actors
over which we have no control.